made math functions use Long and other fixes

+++ src/main/scala/com/gravitydev/scoop/functions.scala
@@ -4,9 +4,9 @@ package functions
 import query._
 
 object `package` {
-  def count (expr: ast.SqlExpr[_]) = sql[Int]("COUNT(" +~ expr +~ ")")
-  def countDistinct (expr: ast.SqlExpr[_]) = sql[Int]("COUNT(DISTINCT " +~ expr +~ ")")
-  def sum (expr: ast.SqlExpr[_]) = sql[Int]("SUM(" +~ expr +~ ")")
+  def count (expr: ast.SqlExpr[_]) = sql[Long]("COUNT(" +~ expr +~ ")")
+  def countDistinct (expr: ast.SqlExpr[_]) = sql[Long]("COUNT(DISTINCT " +~ expr +~ ")")
+  def sum (expr: ast.SqlExpr[_]) = sql[Long]("SUM(" +~ expr +~ ")")
   def avg [T:SqlType](expr: ast.SqlExpr[T]) = sql[T]("AVG(" +~ expr +~ ")")
   def max [T:SqlType](expr: ast.SqlExpr[T]) = sql[T]("MAX(" +~ expr +~ ")")
   def min [T:SqlType](expr: ast.SqlExpr[T]) = sql[T]("MIN(" +~ expr +~ ")")
